tests/requirements.txt
coverage==4.5.1
PyExecJS==1.5.1
mock==2.0.0
nose==1.3.7
responses==0.9.0